#1329 moving a few tests and demo classes to matching packages
